requirement.txt
django==3.1.7
djangorestframework==3.12.2
mysqlclient==2.0.3
django-rest-knox==4.1.0
graphene-django==2.15.0
PyJWT==1.7.1
django-graphql-jwt==0.3.1